Generate a password if the bridge operator didn't.
Generally, we require a password (the UniformDH shared secret) which is set by
the bridge operator using "ServerTransportOptions" in Tor's torrc.  Since some
bridge operators won't bother taking care of that, we want a fallback mechanism
which silently generates and publishes a password if ServerTransportOptions is
not used.  We (ab)use obfsproxy's get_public_server_options() which is meant to
sanitise parameters.  We, however, use it to add the password parameter if it's
not there.