Help operators learn their server descriptor.
At this point, it is difficult for bridge operators to learn the password if
they did not specify it explicitly.  To make that easier, we always dump the
server descriptor to a file.  As a result, bridge operators could simply look
at that file to get a copy & pastable server descriptor.

This should fix bug <https://bugs.torproject.org/10887>.